+++ WooCommerce/WooCommerceUITests/Tests/LoginTests.swift
@@ -35,15 +35,10 @@ class LoginTests: XCTestCase {
         //verify tapping "Log in" will show options to enter a WordPress.com email, log in with Google, or enter your site address.
         WelcomeScreen()
         .selectLogin()
-        .openHelpMenu()
-        //will add .sendEmailToSupport once I have some questions answered about that
-        .closeHelpMenu()
-
+            
         //verify all login options exist
-        .emailLoginOption()
-        .siteAddressLoginOption()
-        .googleLoginOption()
-        .backToWelcomeScreen()
+        .doAllLoginOptionsExist()
+        .goBackToWelcomeScreen()
 
         XCTAssert(WelcomeScreen.isLoaded())
